Create Report or monitor files comming via FTP by using sql table
This Powershell script allows you to insert filenames and other file properites to a sql server table. I use the script for ftp production to make a report for the users .   you can connect a SSRS report to CORE_FILES table so your end users can see what happened with the files

This Powershell script allows you to insert filenames and other file properites to a sql server table. I use the script for ftp production to make a report for the users .  

you can connect a SSRS report to CORE_FILES table so your end users can see what happened with the files from a reporting environment without the need of loging to a server and searching through a network folder.   
you can schedule the script in sql server agent as powershell type of step , put the code in and run it on schedule.   
 

The parameters speak for themselves:

$FileLocation='\\dbs18\SFTP_Root\'
$Server='DBS18'
$Database='Anwb_Rhl'
$Table='CORE_FILES'
set-Location C:
Set-Location $FileLocation
I added set-location commando to make networkdirectory working from sql server agent. otherwise there was an error and agent could not get into the folder.


save it as .ps1 file