Create Multiday transactions to duration per day adapter.sql
Consider a situation where you have transactional data where each row is a transaction that spans one or more days with begintime and endtime. In this situation you can do calculations and show it either on the begin of the transaction date or the enddate. This is fine quick approach if you focus on totals/cumulative. However this appoach will give inacurate results in between the dates. 

Consider a small example. 
1-1-2019   3-1-2019 Transport  CAR1
2-1-2019   5--1-2019 Transport CAR2
Based on this data we cannot simply chart cars on transport for the second of januari. Hour data is not per day. We would like something like this:
1-1-2019 1 Transport CAR1
1-2-2019 1 Transport CAR1
1-3-2019 1 Transport CAR1
and similar for CAR2. 
This approach allows us easily count transactions per date. Basically for each transaction we create a new number of records equal to number of days between begin anbd end time of the transaction. 
Notice that if we probably have datetimestampls so we should not count the whole day for the duration for the first date but use some logic to deal with the timestamps. 
This is exacly what the adapter is doing: transactions in, duration per day out. 
An advantage of this approach is that it is super fast, the fastest what you can get and it is rather generic.